running - optimized find_triangles_sphere so that it finds 20 triangles instead of 120 and reworked zombeis to use new triangle system and made floor

+++ World.h
@@ -22,12 +22,12 @@ class World {
         int center_z;
         int gravity_dist;
 
-        std::vector<float> surface_color;
+        std::vector<double> surface_color;
 
         std::vector<std::vector<double>> sphere_points_3D;
         std::vector<Triangle> sphere_triangles_3D;
 
-        World(Graphics* graphics, int x, int y, int z, std::vector<float> surface_color); // Constructor
+        World(Graphics* graphics, int x, int y, int z, std::vector<double> surface_color); // Constructor
         
 };
